Extract AbstractRenderer component
